pdfApp/requirements.txt
flask==3.0.3
pypdf==4.3.1
Pillow==10.4.0
PyMuPDF==1.24.9